Move SDK flag getters to `utilities`
