+++ src/utilities/utilities.ts
@@ -101,6 +101,29 @@ export async function execSwift(
 }
 
 /**
+ * Get SDK flags for SwiftPM
+ */
+export function swiftpmSDKFlags(): string[] {
+    if (configuration.sdk.length > 0) {
+        return ["--sdk", configuration.sdk];
+    }
+    return [];
+}
+
+/**
+ * Get SDK flags for swiftc
+ *
+ * @param indirect whether to pass the flags by -Xswiftc
+ */
+export function swiftDriverSDKFlags(indirect = false): string[] {
+    if (configuration.sdk !== "") {
+        return [];
+    }
+    const args = ["-sdk", configuration.sdk];
+    return indirect ? args.flatMap(arg => ["-Xswiftc", arg]) : args;
+}
+
+/**
  * Get the file name of executable
  *
  * @param exe name of executable to return
